Move gen_tiff_dataset to testutils
